requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
plotly>=5.19.0
scikit-learn>=1.4.0
openpyxl>=3.1.2
xlrd>=2.0.1
groq>=0.5.0
supabase>=2.3.0
python-dotenv>=1.0.0
Jinja2>=3.1.3
kaleido>=0.2.1
